+++ projects/oculr-ngx/src/lib/services/event-dispatch.service.spec.ts
@@ -53,58 +53,6 @@ describe('EventDispatchService', () => {
     });
   });
 
-  describe('trackChange', () => {
-    it('dispatches a CHANGE_EVENT event with model and location', () => {
-      const mockEvent: any = { id: 'mock' };
-      const mockEventDispatch: any = {
-        id: 'mock',
-        eventType: AnalyticEventType.CHANGE_EVENT,
-        location: mockLocation,
-      };
-      eventDispatchService.trackChange(mockEvent);
-      expect(mockEventBus.dispatch).toHaveBeenCalledWith(mockEventDispatch);
-    });
-  });
-
-  describe('trackClick', () => {
-    it('dispatches a CLICK_EVENT event with model and location', () => {
-      const mockEvent: any = { id: 'mock' };
-      const mockEventDispatch: any = {
-        id: 'mock',
-        eventType: AnalyticEventType.CLICK_EVENT,
-        location: mockLocation,
-      };
-      eventDispatchService.trackClick(mockEvent);
-      expect(mockEventBus.dispatch).toHaveBeenCalledWith(mockEventDispatch);
-    });
-  });
-
-  describe('trackDisplay', () => {
-    it('dispatches a DISPLAY_EVENT event with model and location', () => {
-      const mockEvent: any = { id: 'mock' };
-      const mockEventDispatch: any = {
-        id: 'mock',
-        eventType: AnalyticEventType.DISPLAY_EVENT,
-        location: mockLocation,
-      };
-      eventDispatchService.trackDisplay(mockEvent);
-      expect(mockEventBus.dispatch).toHaveBeenCalledWith(mockEventDispatch);
-    });
-  });
-
-  describe('trackFocus', () => {
-    it('dispatches a FOCUS_EVENT event with model and location', () => {
-      const mockEvent: any = { id: 'mock' };
-      const mockEventDispatch: any = {
-        id: 'mock',
-        eventType: AnalyticEventType.FOCUS_EVENT,
-        location: mockLocation,
-      };
-      eventDispatchService.trackFocus(mockEvent);
-      expect(mockEventBus.dispatch).toHaveBeenCalledWith(mockEventDispatch);
-    });
-  });
-
   describe('trackAppEvent', () => {
     it('dispatches a APP_EVENT event with model and location', () => {
       const mockEvent: any = { id: 'app-init' };
@@ -128,86 +76,6 @@ describe('EventDispatchService', () => {
       };
       eventDispatchService.trackAppError(mockEvent);
       expect(mockEventBus.dispatch).toHaveBeenCalledWith(mockEventDispatch);
-    });
-  });
-
-  describe('trackValidationError', () => {
-    it('dispatches a VALIDATION_ERROR_EVENT event with model and location', () => {
-      const mockEvent: any = { id: 'error' };
-      const mockEventDispatch: any = {
-        id: 'error',
-        eventType: AnalyticEventType.VALIDATION_ERROR_EVENT,
-        location: mockLocation,
-      };
-      eventDispatchService.trackValidationError(mockEvent);
-      expect(mockEventBus.dispatch).toHaveBeenCalledWith(mockEventDispatch);
-    });
-  });
-
-  describe('API Event Actions', () => {
-    const httpRequest: any = { url: 'http://localhost:3000/breed/getAll', method: 'GET' };
-    const httpResponse: any = { url: 'http://localhost:3000/breed/getAll', status: 200 };
-
-    describe('trackApiStart', () => {
-      it('dispatches an API_START_EVENT with the supplied id and scopes', () => {
-        const apiContext: any = { id: 'api start', scopes: [{ data: 'starting' }] };
-        const expectedEvent: any = {
-          request: httpRequest,
-          id: apiContext.id,
-          scopes: apiContext.scopes,
-          eventType: AnalyticEventType.API_START_EVENT,
-          location: mockLocation,
-        };
-        eventDispatchService.trackApiStart(apiContext, httpRequest);
-        expect(mockEventBus.dispatch).toHaveBeenCalledWith(expectedEvent);
-      });
-
-      it('dispatches an API_START_EVENT with the api url as an id', () => {
-        const apiContext: any = {};
-        const expectedEvent: any = {
-          request: httpRequest,
-          id: httpRequest.url,
-          scopes: [],
-          eventType: AnalyticEventType.API_START_EVENT,
-          location: mockLocation,
-        };
-        eventDispatchService.trackApiStart(apiContext, httpRequest);
-        expect(mockEventBus.dispatch).toHaveBeenCalledWith(expectedEvent);
-      });
-    });
-
-    describe('trackApiComplete', () => {
-      it('dispatches an API_COMPLETE_EVENT with the supplied id and scopes', () => {
-        const apiContext: any = { id: 'api complete', scopes: [{ data: 'starting' }] };
-        const duration = 312;
-        const expectedEvent: any = {
-          response: httpResponse,
-          request: httpRequest,
-          duration,
-          id: apiContext.id,
-          scopes: apiContext.scopes,
-          eventType: AnalyticEventType.API_COMPLETE_EVENT,
-          location: mockLocation,
-        };
-        eventDispatchService.trackApiComplete(apiContext, httpResponse, httpRequest, duration);
-        expect(mockEventBus.dispatch).toHaveBeenCalledWith(expectedEvent);
-      });
-
-      it('dispatches an API_COMPLETE_EVENT with the api url as an id', () => {
-        const apiContext: any = {};
-        const duration = 312;
-        const expectedEvent: any = {
-          response: httpResponse,
-          request: httpRequest,
-          duration,
-          id: httpRequest.url,
-          scopes: [],
-          eventType: AnalyticEventType.API_COMPLETE_EVENT,
-          location: mockLocation,
-        };
-        eventDispatchService.trackApiComplete(apiContext, httpResponse, httpRequest, duration);
-        expect(mockEventBus.dispatch).toHaveBeenCalledWith(expectedEvent);
-      });
     });
   });
 });
